5758 - Refactor and move descriptions utility
