+++ src/meta/assessment/descriptions/index.ts
@@ -1,0 +1,9 @@
+import { getLabelKey } from 'meta/assessment/descriptions/getLabelKey'
+import { getVisibleDescriptionNames } from 'meta/assessment/descriptions/getVisibleDescriptionNames'
+import { hasSectionOnlyComments } from 'meta/assessment/descriptions/hasSectionOnlyComments'
+
+export const Descriptions = {
+  getLabelKey,
+  getVisibleDescriptionNames,
+  hasSectionOnlyComments,
+}
