+++ src/meta/assessment/descriptions/types.ts
@@ -1,0 +1,7 @@
+import { SectionName } from 'meta/assessment/section'
+
+export type DescriptionVisibilityProps = {
+  hasNationalDataPointData: boolean
+  sectionName: SectionName
+  useNationalDataPoint: boolean
+}
