+++ src/meta/cycleData/links/getLocationLabel.ts
@@ -3,7 +3,7 @@ import { TFunction } from 'i18next'
 import { CountryIso } from 'meta/area/countryIso'
 import { Assessment } from 'meta/assessment/assessment'
 import { Cycle } from 'meta/assessment/cycle'
-import { Descriptions } from 'meta/assessment/description/descriptions'
+import { Descriptions } from 'meta/assessment/descriptions'
 import { CommentableDescriptionName } from 'meta/assessment/descriptionValue'
 import { Labels } from 'meta/assessment/labels'
 import { SubSection } from 'meta/assessment/section'
